Add test module and new test
